Bigger color framebuffer

+++ Atlas/src/Atlas/Renderer/Renderer.cpp
@@ -184,7 +184,7 @@ namespace Atlas
 	{
 		FramebufferSpecification fbSpec;
 		// Color, EntityID, PostProcessing, Depth
-		fbSpec.Attachments = { FramebufferTextureFormat::RGBA8, FramebufferTextureFormat::RED_INTEGER, FramebufferTextureFormat::RGBA8, FramebufferTextureFormat::Depth };
+		fbSpec.Attachments = { FramebufferTextureFormat::RGBA16, FramebufferTextureFormat::RED_INTEGER, FramebufferTextureFormat::RGBA16, FramebufferTextureFormat::Depth };
 		fbSpec.Width = 1280;
 		fbSpec.Height = 720;
 		s_RendererData.Framebuffer = Framebuffer::Create(fbSpec);
